커스텀 명령어 탭 내 레이아웃 개선

+++ src/containers/CustomContainer.jsx
@@ -1,12 +1,15 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 
+import ListLayout from "../layouts/ListLayout";
 import CustomAddButton from "../presentationals/CustomAddButton";
 import CustomForm from "../presentationals/CustomForm";
-import CustomList from "../presentationals/CustomList";
+import CustomItem from "../presentationals/CustomItem";
+import SideBarHeader from "../presentationals/SideBarHeader";
+import EmptyItem from "../presentationals/EmptyItem";
 import { deleteCustomCommand, setCustomCommandList, setCustomFormLatex, setCustomFormValue } from "../slice";
 
-export default function CustomContainer() {
+export default function CustomContainer({ onScroll }) {
 	const { customCommandList, customFormValue } = useSelector(state => state);
 	const dispatch = useDispatch();
 
@@ -21,9 +24,11 @@ export default function CustomContainer() {
 	};
 
 	const handleDeleteClick = index => () => {
-		const tempCustomCommands = [...customCommandList].filter((_, id) => id !== index);
+		if (confirm("정말로 삭제하시겠습니까?")) {
+			const tempCustomCommands = [...customCommandList].filter((_, id) => id !== index);
 
-		dispatch(deleteCustomCommand({ customFormValue, tempCustomCommands }));
+			dispatch(deleteCustomCommand({ customFormValue, tempCustomCommands }));
+		}
 	};
 
 	const handleSubmit = e => {
@@ -68,7 +73,7 @@ export default function CustomContainer() {
 	};
 
 	return (
-		<div>
+		<ListLayout onScroll={onScroll}>
 			<CustomAddButton
 				isFormOn={customFormValue.state}
 				onClick={handleFormOnButton}
@@ -80,11 +85,17 @@ export default function CustomContainer() {
 					onChangeLatex={onChangeLatex}
 					onSubmit={handleSubmit}
 				/>}
-			<CustomList
-				customs={customCommandList}
-				onClickEdit={handleEditClick}
-				onClickDelete={handleDeleteClick}
-			/>
-		</div>
+			<SideBarHeader title={"사용자 명령어 목록"} />
+			{customCommandList.length ?
+				customCommandList.map(({ command }, index) =>
+					<CustomItem
+						key={index}
+						name={command}
+						onClickEdit={handleEditClick(index)}
+						onClickDelete={handleDeleteClick(index)}
+					/>,
+				) :
+				<EmptyItem content={"최근 저장한 명령어가 없습니다"}/>}
+		</ListLayout>
 	);
 }
